==> code/01 combine_weekly2019.ipynb <==
# Imports 
import pandas as pd

Query used to get weekly data from [Stackoverflow Query](https://data.stackexchange.com/stackoverflow/query/edit/1205218)

```SQL
SELECT Id, CreationDate, Title, Body, Tags, ViewCount, AnswerCount, CommentCount
FROM Posts
WHERE (CreationDate BETWEEN '2019-01-14' AND '2019-01-20') AND Tags IS NOT NULL
ORDER BY CreationDate ASC
```

# Notebook Overview: 
- Combine weekly .csv's into a combined .csv for the year of 2019 
- The following are links to the zipped versions of the weekly .csv files:
    - [Weeks 1 - 20](https://drive.google.com/open?id=1uttbh17hfhnfMLqnajbw3yX7LC27yQFY)
    - [Weeks 21-40](https://drive.google.com/open?id=1Isrpl4XX-Sv3CDmI2eNI26fX8Hu98mt4)
    - [Weeks 41-52](https://drive.google.com/open?id=15FEnPmx_LaP9_xpxB3zz6BMqFDgTbVc6)
- [Combined 2019 data (zipped):](https://drive.google.com/open?id=1ZQHpysA5m8aL3bbMEiUJLdHmVDhZmaBq)
 

## Function

def combine_weekly(n_weeks = 52):
    '''
    Function to combine weekly .csv files into 
    one combined .csv file
    
    Note: For the function to work you will need to download the  
    individual .csv files (linked above) and follow the listed 
    file path: 
    '../data/individual_weekly/[week#.csv]'
    
    '''
    start = 1 
    weekly_list = []
    
    for i in range(start, n_weeks + 1):
        week_i = pd.read_csv(f'../data/individual_weekly/week{i}.csv')
        weekly_list.append(week_i)
    
    combined_csv = pd.concat(weekly_list)
    return combined_csv  

combined_csv = combine_weekly(n_weeks = 52)

combined_csv.head()

combined_csv.tail()

combined_csv.shape

# Save combined csv to data folder 
# will use this file in the next notbook: 02 preprocessing_cleaning 
combined_csv.to_csv('../data/weekly_combined2019.csv', index=False)
