==> code/02 preprocessing_cleaning .ipynb <==
# Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

import seaborn as sns
from datetime import datetime

import regex as re
from bs4 import BeautifulSoup 

# Data Cleaning: 
1. Language tags 
    - Target Tags: 
        - SQL (Target Class: 0)
        - Scala (Target Class: 1)
        - R (Target Class: 2)
        - Julia (Target Class: 3)
        - C++ (Target Class: 4)
        - Python (Target Class: 5)
    - Remove tags that won't be used for modeling (the 6 above) 
2. Body & Title (will I use Title in modeling?) 
    - Clean up body and titles:
        - remove HTML tags & numbers 
3. MISC Cleaning
    - Reset Index
    - CreationDate column --> Datetimeformat

#### To continue with this notebook you will need to have weekly_combined2019.csv file. [link](https://drive.google.com/open?id=1OPgd2J14mfQkCpvhz9O8qFbql0FUpDaf)

df = pd.read_csv('../data/weekly_combined2019.csv')

df.head()

## Language Tags 

### Clean up tags

def tag_replace_split(tag):
    '''
    function to get rid of characters ('><, <, >') from tag column and replace with either space, or nothing 
    then put that new tag into list to be able to parse more easily for modeling  

    '''
    tag = tag.replace('><', ' ')
    tag = tag.replace('<', '')
    tag = tag.replace('>', '')
    tag = tag.split()
    
    return tag 

# Use "tag_replace_split" function to clean up original tags column  
df['Tags'] = df['Tags'].map(lambda x: tag_replace_split(x))

df.head()

### Categorize Langauge Tags
- categorizing and drop unnecesary rows with non-target tags

#### Language Tag Sets

# Individual tag sets for each language I'm targeting 
# created target sets

# SQL (target class: 0)
sql_set = {'sql','bigquery-standard-sql','sqlparameters', 'sqlparameter', 'web-sql', 'sqlite3','sqlite', 
          'sql2016', 'sql-server-2016', 'sql-execution-plan','spark-sql','sql-order-by','sql-join',  
          'sql-syntax','sql-statement','sql-query','mysql-query','orderby','sql-server','mysql'} 

# Scala (taget class: 1)
scala_set = {'scala', 'scala-implicits', 'scala-generics'} 

# R (target class: 2)
r_set = {'rstats', 'r', 'r-language', 'r-faq'}

# Julia (target: 3)
julia_set = {'julia-lang', 'julia', 'julialang'}

# C++ (target: 4)
cplus_set = {'c++', 'c++17', 'visual-c++'}

# Python (target: 5)
py_set = {'python', 'python3', 'python-3.7', 'python-3.x', 'pytorch', 'python-list', 'py3', 'ipython-notebook', 
          'python3', 'python-3.5.2' , 'python-2.7', 'python-3', 'python-bigquery', 'pythonic', 'py'}

# includes all language tags 
# union of all sets 
master_set = (sql_set|scala_set|r_set|julia_set|cplus_set|py_set)

# else : -1 

all_sets = [sql_set, scala_set, r_set, julia_set, cplus_set, py_set]

# look through each row of tags and see if there's an intersection with tags and master set of languages 
# has to be a set to use .intersection 
df['target_tags'] = df['Tags'].apply(lambda tags: set(tags).intersection(master_set))

df.head()

# Original 
df.shape

# drop rows that have empty sets for target_tags 
df.drop(df[df['target_tags'] == set({})].index, inplace = True)

# New df with only target languages from sets created 
# loss about 75% of data from original
df.shape

# difference between old and new / old 
((1878570-465055)/1878570)*100

#### Overlapping Target Classes

# use this column to check and see which languages are overlapping b/c can't use .unique on list 
df['overlap_tags'] = df['target_tags'].map(lambda tag: ' '.join(tag))

# overlapping languagues: 
# Fair amount, will need to address this. 
# r & python 
# sql & python 
# r & sql
# c ++ & sql 
# 'python-3.x scala sql-server'
# 'python-3.x python-2.7 c++'
# 'python-3.x python r sql'

df['overlap_tags'].unique()

# If have time will come back and create functions to check for overlapping classes 
def check_2class(tags, class1, class2):
    '''
    make a function that would take a post's tags, and check 
    if the tags are in class1 and class2.
    
    If so, return the rows that satisfy that condition.
    
    (Then you can drop those rows --> do this outside the function)
    '''
    pass

def check_3class(tags, class1, class2, class3):
    pass

def check_4class(tags, class1, class2, class3, class4):
    pass

def check_5class(tags, class1, class2, class3, class4, class5):
    pass

def check_6class(tags, class1, class2, class3, class4, class5, class6):
    pass

df.head()

#### Create Target Classes

- 0 = SQL  
- 1 = Scala  
- 2 = R 
- 3 = Julia  
- 4 = C++  
- 5 = Python 

def assign_target(tags):
    """
    Function to create target classes for my tags 
    
    *Limitation*: uses first tag to assign the target. 
    If there are multiple langauges in the tags like SQL and Python, 
    will pick SQL since appears first. 
    
    """
    for target_class, target_set in enumerate(all_sets):
        for tag in tags.split():
            if tag in target_set:
                return target_class
            else:
                continue
    return -1

df['target_class'] =  df['overlap_tags'].map(assign_target)
df.head()

# no -1, so there aren't any extra classes 
df['target_class'].unique()

#### Target Class Baseline 

df['target_class'].value_counts(normalize = True)

## Body Cleanup
    - future iterations of the project, will consider adding titles as well. 

def body_cleanup(raw_body):
    '''
    Function to convert original body to a string of words: 
    - remove HTML artifacts & numbers
    - convert to lowercase then join as one string
    '''
    
    # Remove HTML artifacts using BeautifulSoup 
    body_text = BeautifulSoup(raw_body).get_text()
    
    # Remove non-letters (returns only letters) -- Regex 
    letters_only = re.sub("[^a-zA-Z]", " ", body_text)
    
    # Convert to lower case, split into individual words (tokenize)
    words = letters_only.lower().split()
    
    # Join the words back into one string separated by space, and return the result.
    return(" ".join(words))

# Execution time: ~5m 56s 
df['body_clean'] = df['Body'].map(lambda x: body_cleanup(x))

df.head()

## MISC Cleaning 

### Reset Index

df.reset_index(inplace = True)

df.head()

# Drop 'index' column (b/c I reset index above)
df = df.drop(['index'], axis = 1)

### Check for nulls

df.isnull().sum()

# In EDA notebook, find null, because body_clean is empty
# in above cleaning function, need to add: Nan if empty -- can fix for next iteration of project 
# df[df['Id'] == 55234780]

### Check Dtypes

df.dtypes

#### Datetime formatt 

df['CreationDate']= pd.to_datetime(df['CreationDate']) 

# Save as clean_final.csv

df.to_csv('../data/clean_final.csv', index=False)
