==> code/04 modeling.ipynb <==
# Imports

import pandas as pd
import numpy as np 

from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.pipeline import Pipeline

from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer


# Which stop words do I want to use? -- for this project will use native sklearns 
from sklearn.feature_extraction import stop_words

# import nltk
# from nltk.corpus import stopwords

np.random.seed(42)

df = pd.read_csv('../data/clean_final.csv')
df.head()

df.isnull().sum()

df[df['body_clean'].isnull()]

df.loc[91170, 'Body']

# drop row with missing value in 'body_clean' (technically any missing values)
df = df.dropna()

df.isnull().sum()

# Modeling 

## Model Prep 

X = df['body_clean']
y = df['target_class']

Target Tags: 
- 0 =  SQL
- 1 = Scala
- 2 = R
- 3 = Julia
- 4 = C++
- 5 = Python

### Basline Accuracies 

y.value_counts(normalize = True)

X_train,X_test, y_train, y_test = train_test_split(X,
                                                   y,
                                                   test_size = 0.25,
                                                   stratify = y,
                                                   random_state = 42)

### CountVectorizer + Logistic Regression: 
#### Create Pipeline and Gridsearch

# CountVectorizer Pipeline:
# 1. CountVectorizer Instance (transformer)
# 2. LogisticRegression Instance (estimator)

max_iterations = 2_000

pipe_cvec = Pipeline([
    ('cvec',CountVectorizer()),        
    ('lr',LogisticRegression(max_iter = max_iterations,           # To help mitigate convergence warnings
                             solver = 'lbfgs',                    # turns off future warnings 
                             multi_class = 'multinomial'))        # mitigates future warnings  
])

# param grid 

pipe_cvec_params = {
    'cvec__max_features': [400,600],           # was initially [100, 500]  --> increased [400,600]
    'cvec__ngram_range': [(1,2)],              # got rid of (1,1) -- only using (1,2) for now 
    'cvec__stop_words' : ['english']           # can consider adding to stop words dictionary 
    # can decide if I want to use these hyperparameters 'max_df=1.0' and 'min_df=1' for CVEC params 
    
}

# Instantiate CountVectorizer GridSearchCV
gs_cvec = GridSearchCV(pipe_cvec,
                       param_grid = pipe_cvec_params,
                       cv = 3)

# Fit GridSearch(CVEC) to training data.
gs_cvec.fit(X_train,y_train)

gs_cvec.best_params_

# save best model 
gs_cvec_model = gs_cvec.best_estimator_

# CountVectorize Accuracy on training 
gs_cvec_model.score(X_train,y_train)

# CountVectorize Accuracy on testing
gs_cvec_model.score(X_test,y_test)

# Training - Testing 
0.8996387511109837- 0.8967608202022982

### TFIDFVectorizer + Logistic Regression: 
#### Create Pipeline and Gridsearch

# TFIDF Vectorizer Pipeline:
# 1. TFIDF Instance (transformer)
# 2. LogisticRegression Instance (estimator)

max_iterations = 2_000

pipe_tvec = Pipeline([
    ('tvec',TfidfVectorizer()),        
    ('lr',LogisticRegression(max_iter = max_iterations,           # To help mitigate convergence warnings
                             solver = 'lbfgs',                    # turns off future warnings 
                             multi_class = 'multinomial'))        # mitigates future warnings  
])

# max_features = [800,1000] --> decreased accuracy 
# TFIDF param grid 

pipe_tvec_params = {
    'tvec__max_features': [600, 800],          # was initially [100, 500]  --> increased [400,600] --> [600,800]--> [800,1000] 
    'tvec__ngram_range': [(1,2)],              # got rid of (1,1) -- only using (1,2) for now 
    'tvec__stop_words' : ['english']           # can consider adding to stop words dictionary 
}

# Instantiate TFIDF GridSearchCV
gs_tvec = GridSearchCV(pipe_tvec,
                       param_grid = pipe_tvec_params,
                       cv = 3)

# Fit GridSearch(TVEC) to training data.
gs_tvec.fit(X_train,y_train)

gs_tvec.best_params_

# save best model 
gs_tvec_model = gs_tvec.best_estimator_

# TFIDF Accuracy on training 
gs_tvec_model.score(X_train,y_train)

# TFIDF Accuracy on testing
gs_tvec_model.score(X_test,y_test)

0.9142865334441928 - 0.9124750567673571

#### Results -- Best
- Best model: 
    - Highest accuracy
    - Not overfit 

### TFIDFVectorizer + Decision Trees: 
- Can look into better hyperparameters
#### Create Pipeline and Gridsearch

# TFIDF Vectorizer Pipeline:
# 1. TFIDF Instance (transformer)
# 2. Decision Tree Classifer (estimator)

rt_pipe_tvec = Pipeline([
                    ('tvec', TfidfVectorizer()),     
                    ('rt', DecisionTreeClassifier(random_state = 42))
                                 
])

# Setting up Decision Tree (rt) & tvec params
rt_tvec_params = {
    'tvec__max_features':[600],              # 
    'tvec__ngram_range': [(1, 2)],
    'tvec__stop_words': ['english'],
    'rt__max_depth': [5,10],                 # [3,4,5,6,7]
    'rt__min_samples_split': [7,15],         # [7,15,30]  
    'rt__min_samples_leaf': [3,10]           # [ 3, 4, 5,6] 
 
}

# Decision Tree and TVEC Gridsearch
gs_rt_tvec = GridSearchCV(rt_pipe_tvec, 
                          param_grid = rt_tvec_params,
                          cv = 3)

gs_rt_tvec.fit(X_train,y_train)

gs_rt_tvec.best_params_

# save best model 
gs_rt_tvec_model = gs_rt_tvec.best_estimator_

# Accuracy on training 
gs_rt_tvec_model.score(X_train,y_train)

# Accuracy on testing
gs_rt_tvec_model.score(X_test,y_test)

0.829909114366811 - 0.8246060689465354

#### Results
- Not overfit, but accuracy isn't as high as the TFDIF-Logisitc Regression model

# Future Modeling 

### TFIDFVectorizer + Random Forest: 
- Need to revisit this in the future:
    - fitting a random forest is taking too long to fit
        - Since I am going to revist Random Forest modeling, I will keep in my notebook for future use. 

#### Create Pipeline and Gridsearch

# TFIDF Vectorizer Pipeline:
# 1. TFIDF Instance (transformer)
# 2. Random Forest Classifier Instance (estimator)

rf_pipe_tvec = Pipeline([
                    ('tvec', TfidfVectorizer()),        
                    ('rf', RandomForestClassifier(random_state = 42))
                                 
])

# If the above doesn't work try this way to set up params:
# Setting up randomforest & tvec params
rf_tvec_params = {
    'tvec__max_features':[600],                      # vary max features [600,800,1000]
    'tvec__ngram_range': [(1, 2)],                
    'tvec__stop_words': ['english'],
    'rf__n_estimators': [100], # [5,6,10,100]  
    'rf__max_depth': [3,5],      # [3,4,5]
    'rf__max_features': [3, 4] # [None, 3, 4]  , None = Bagging 
 

#  'rf__min_samples_split': [100],
#  'rf__max_leaf_nodes': [None]
}

# # Random Forest and TVEC Gridsearch
gs_rf_tvec = GridSearchCV(rf_pipe_tvec, 
                          param_grid = rf_tvec_params,
                          cv = 3)

# Fit GridSearch to training data.
# TOOK OVER 3 hours, and still didn't fit...look at params, use less hyper parameters for first run
# see if that speeds up the fitting process. 

gs_rf_tvec.fit(X_train,y_train)

gs_rf_tvec.best_params_

# save best model 
gs_rf_tvec_model = gs_rf_tvec.best_estimator_

# Accuracy on training 
gs_rf_tvec_model.score(X_train,y_train)

# Accuracy on testing
gs_rf_tvec_model.score(X_test,y_test)
